move models to model.dart
